+++ CHANGELOG.md
@@ -65,6 +65,7 @@ This name should be decided amongst the team before the release.
 - [#1085](https://github.com/tweag/topiary/pull/1085) CSS formatting fixes
 - [#1092](https://github.com/tweag/topiary/pull/1092) Log early returns during formatting
 - [#1113](https://github.com/tweag/topiary/pull/1113) Update `dist` to 0.30
+- [#1117](https://github.com/tweag/topiary/pull/1117) Handle inline comments and trailing commas in WIT
 
 ## v0.6.1 - Gilded Ginkgo - 2025-05-27